Refactor environment variables
We have a lot of environment variables that end up being different
URLs on the same domain. Each URL is only used once...so this moves the domain,
the only part that needs to be configurable, into an environment variable
and bakes the URI path into variables that make use of the domain env var.

This also deletes contants that were defined in ./config/environment.rb
but never used.