Refactor error_type_object
Rails's load order is different between RAILS_ENV=development and RAILS_ENV=production.
This was causing  Rails.application.routes.url_helpers to be in scope in development.
In production it would be in scope but at the time of assignment, it wouldn't have the _path helpers
from routes.rb in it yet.